Rename multi commit operation components and classes
